requirements.txt
Django==3.1.2
django-cors-headers==3.7.0
django_nested_admin==3.3.3
djangorestframework_simplejwt==4.7.1
django_heroku==0.3.1
djangorestframework==3.12.4
django-allauth==0.45.0
django-rest-auth==0.9.5
drf-yasg==1.20.0
gunicorn==20.1.0
Pillow==8.2.0
python-decouple==3.4
